# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hawk-Eye Control Plane",
    "description": "Insider & privileged-user fraud detection — FastAPI control plane (ALERT-ONLY).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/alerts/{alert_id}/assign": {
      "post": {
        "tags": [
          "disposition"
        ],
        "summary": "Assign",
        "operationId": "assign_api_v1_alerts__alert_id__assign_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Assign Api V1 Alerts  Alert Id  Assign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/disposition": {
      "post": {
        "tags": [
          "disposition"
        ],
        "summary": "Disposition",
        "operationId": "disposition_api_v1_alerts__alert_id__disposition_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispositionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/block-request": {
      "post": {
        "tags": [
          "disposition"
        ],
        "summary": "Block Request",
        "description": "Raise a block REQUEST (human action). Never auto-executed — Lead must approve downstream.",
        "operationId": "block_request_api_v1_alerts__alert_id__block_request_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/block-request/approve": {
      "post": {
        "tags": [
          "disposition"
        ],
        "summary": "Approve Block Request",
        "description": "AGM Vigilance APPROVES a raised block request (Part 24.1: ✅ approve for Request block).\n\nThis is the human approval of the *request* — it is still **never an auto-block of money**.\nActioning a block remains a manual, out-of-band step; the system only records the approval.",
        "operationId": "approve_block_request_api_v1_alerts__alert_id__block_request_approve_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Approve Block Request Api V1 Alerts  Alert Id  Block Request Approve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignRequest": {
        "properties": {
          "assignee": {
            "type": "string",
            "title": "Assignee",
            "description": "Investigator id to assign/claim",
            "examples": [
              "EMP-aud1"
            ]
          }
        },
        "type": "object",
        "required": [
          "assignee"
        ],
        "title": "AssignRequest"
      },
...
      "BlockRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Why a block is being requested (human action)"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "BlockRequest"
      },
      "BlockRequestResponse": {
        "properties": {
          "alert_id": {
            "type": "string",
            "title": "Alert Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Pending AGM Vigilance approval — never auto-executed",
            "default": "block_requested"
          },
          "requires_approval_by": {
            "type": "string",
            "title": "Requires Approval By",
            "default": "agm_vigilance"
          },
          "auto_blocked": {
            "type": "boolean",
            "title": "Auto Blocked",
            "description": "Always False — the system never auto-blocks money",
            "default": false
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "audit_id"
        ],
        "title": "BlockRequestResponse",
        "description": "A *request*, never an executed block. Requires AGM Vigilance approval to action downstream."
      },
...
      "DispositionOutcome": {
        "type": "string",
        "enum": [
          "fraud",
          "false_positive",
          "inconclusive"
        ],
        "title": "DispositionOutcome"
      },
      "DispositionRequest": {
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DispositionOutcome",
            "examples": [
              "fraud"
            ]
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "description": "Investigator rationale (free text)",
            "default": ""
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids",
            "examples": [
              [
                "evt_8f2a1c90",
                "evt_8f2a1d04"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "outcome"
        ],
        "title": "DispositionRequest"
      },
      "DispositionResponse": {
        "properties": {
          "alert_id": {
            "type": "string",
            "title": "Alert Id",
            "examples": [
              "alr_3d7e22"
            ]
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "confirmed_fraud"
            ]
          },
          "label_written": {
            "type": "boolean",
            "title": "Label Written",
            "default": true
          },
          "feedback_queued_for_retraining": {
            "type": "boolean",
            "title": "Feedback Queued For Retraining",
            "default": true
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id",
            "examples": [
              "aud_99f0c1"
            ]
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "status",
          "audit_id"
        ],
        "title": "DispositionResponse",
        "description": "Exact Part 24.5c response shape."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}